// src/file-manager/google-cloud-storage/interfaces/config.interface.ts
export interface GoogleStorageConfigInterface {
  mediaBucket: string;
  googleAccType: string;
  googleProjectId: string;
  googlePrivateKeyID: string;
  googlePrivateKey: string;
  googleClientEmail: string;
  googleClientId: string;
  googleAuthUri: string;
  googleTokenUri: string;
  googleAuthProviderCertUrl: string;
  googleClientCertUrl: string;
  googleUniverseDomain: string;
  googleCredentials: object;
}